CODEOWNERS: add @carlocab for `universal_binary_allowlist`
Now that we have a `deuniversalize_machos` method available for use in
formulae, we shouldn't be adding new formulae to this allowlist.

I'm adding myself here to make sure I can chime in when someone tries
to, but also so that I can more quickly spot problems with the
`deuniversalize_machos` method, if any.