samba: apply "Fixes the Grouplimit of 16 users os OS X." patch
This patch fixes `get_user_groups: failed to get the unix group list` error
that occurs when the user belongs to more than 16 groups.

The patch is from https://github.com/Homebrew/legacy-homebrew/pull/10311/files .
Discussed in https://bugzilla.samba.org/show_bug.cgi?id=8773 .

This patch is required for running Samba on GitHub Actions with the default
macOS 10.15 runner, which assigns 18 groups for the "runner" user.

```console
$ groups
staff wheel everyone localaccounts _appserverusr admin _appserveradm _lpadmin
_webdeveloper com.apple.access_ssh com.apple.sharepoint.group.1 _appstore
_lpoperator _developer _analyticsusers com.apple.access_ftp
com.apple.access_screensharing com.apple.access_remote_ae
```

The custom GHA runners used in the Homebrew project are not affected.

Other than applying this patch, there is no known workaround to avoid the error.

Note that just dropping the supplemental groups of the samba-dot-org-smbd
process (with a helper tool that calls setgroup(2) syscall) does not seem to
work as a workaround, because samba-dot-org-smbd calls getgrouplist(3) which
talks to opendirectoryd(8) to obtain the group list, regardless to the current
groups of the samba-dot-org-smbd process.
https://github.com/samba-team/samba/blob/samba-4.14.7/source3/lib/system_smbd.c#L185

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>