spdx-sbom-generator 0.0.13 (new formula)
This commit adds Formula/spdx-sbom-generator.rb that includes the required
functionality to build and install the spdx-sbom-generator cli tool.

fix: `brew audit --new spdx-sbom-generator` errors resolved

Split the offending line (16) and turned the regex into a standard
string to avoid issues with abigouity from the linter.

feat: add check for processor family to build correct arch on amd64

This commit adds the ability to build native binaries for both
x86_64 and amd64 processor types.

feat: new test that runs the whole application end to end

The test downloads a git repo and checks out a specific tag so that
the output will be consistant. It then runs the spdx-sbom-generator
and checks to see if it's output file is created.

fix: resolve issue with ubuntu build failing

Change 1: Added a check for OS so I could build a linux version
Change 2: Changed the github url to use HTTPS to fix ubuntu build

fix: add proper version of python for the project that is downloaded

Change to set the required version of python to 3.8

fix: set python to install as user

Added —user to the command for pip install pipenv

change: instead of pulling from git use a resource

Using a resource `wyzeapy` to download the tar of the source
instead of doing a git clone and checkout.

Update Formula/spdx-sbom-generator.rb

Modify the logic in the install to be cleaner and easier to read

Co-authored-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>

fix: trailing whitespace...

Fixing trailing whitespace

Fix it again...

I think this makes it work! I just have to set the right python in the path!

Why not just make it check itself?

Remove python from test dependencies

Add `go` to the test suite

feat: use go mod init to create empty project

Testing with an empty project and checking for the error is quicker
and more efficient than downloading a sperate project.

Fixes: https://github.com/Homebrew/homebrew-core/pull/84924#discussion_r722050652