librist 0.2.6 (new formula)
Proposed formula for RIST, the "Reliable Internet Stream Transport".
According to the authors, RIST is a "new Specification for multi-vendor
interoperable professional video delivery over the Internet". RIST will
be used by VLC and ffmpeg in their current or next versions.

The code is hosted at https://code.videolan.org/rist

This formula installs the RIST library, its header files and a few basic
test tools which are provided with the library.