subversion: add perl bindings and tests back on arm
Perl bindings were removed from subversion on Apple Silicon machines
in #67531 due to test failures.  However, the bindings being missing
prevents other things that depend on them, such as git-svn, from working
properly on this architecture (see some of the comments on #52490).  The
tests have been updated since that PR, so it appears to be safe to add
the bindings back now without reintroducing the failures.