tsduck 3.28-2551 (new formula)
TSDuck (https://tsduck.io/) is the MPEG Transport Stream Toolkit,
a general-purpose toolbox for Digital TV Engineers working on
transport streams (MPEG-TS). TSDuck is an extensible set of
command-line tools, plugins and development libraries in C++,
Python and Java.

TSDuck is available on macOS, Linux and Windows. Until now, it has
been available on macOS through Homebrew using a third-party tap
("brew tap tsduck/tsduck") since September 2017. Given the maturity
of the project and activity around it, it seems appropriate to move
it to Homebrew core.