ninja: make Python a build dependency
Ninja doesn't actually need Python to build anything. It uses Python
only for generating graphs, so there's no point requiring everyone to
install Python as well.

This is useful for users with non-default prefixes, since this is
relocatable but Python is not. It's also useful for CI, because formulae
that use Ninja to build will no longer need to install Python as well
unless they depend on Python in some other way.

Also, remove the Curl dependency, as the test probably still works with
system Curl.

While we're here, let's install some more useful files from the tarball.