six: build an `all` bottle
We do this by dropping the build for `python@3.7`. No formulae in
Homebrew/core need `six` built for `python@3.7`. Having an `all` bottle
guarantees that no user has to build this formula from source, even if
they are on an old version of macOS.

I've omitted the revision bump; I don't think there's a need to force a
reinstall for this change.