nvm: NVM_DIR is default to $HOME/.nvm
## What
The default NVM_DIR should be $HOME/.nvm not to destroy user-installed nodes on upgrades.
nvm.sh and nvm-exec are expected to be in NVM_DIR so that the shim script symlinks them into NVM_DIR.
This would reduce bug reports to upstream.
https://github.com/Homebrew/homebrew-core/pull/86702#issuecomment-937170891

## How
The existing `nvm.sh` is replaced with a shim script and `nvm.sh` from upstream is moved to `libexec`.
The shim script is basically loading `nvm.sh` with some additional operation to make things safer.
What the shim script does is as follows:
* If `$NVM_DIR` is not set, set it to `$HOME/.nvm`, which the upstream recommends as `$NVM_DIR`.
* Load `nvm.sh` placed in libexec. Note that this shim script will be generated every version so the path includes Cellar. Backslash in front of `.` is to avoid `alias .=something`
* If `$NVM_DIR` does not exist, create the directory on `$NVM_DIR`.
* If `$NVM_DIR/nvm.sh` does not exist, create a symlink at `$NVM_DIR/nvm.sh` to the Homebrew-provided `nvm.sh`. The upstream recommends nvm.sh should be under $NVM_DIR. Note that this symlink will be created only for the first time so that the link uses opt_prefix.
* If `$NVM_DIR/nvm-exec` does not exist, create a symlink at `$NVM_DIR/nvm-exec` to the Homebrew-provided `nvm.sh`. The upstream recommends nvm-exec should be under $NVM_DIR. Note that this symlink will be created only for the first time so that the link uses opt_prefix.

For those who set `$NVM_DIR` to `/usr/local/opt/nvm`, create a symlink at `#{prefix}/nvm-exec` to `#{libexec}/nvm-exec`.

Co-authored-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>