ranger: add missing command 'rifle'
Define an entry for command 'rifle'.

Running the `rifle.py` in `#{libexec}/ranger/ext/rifle.py` doesn't work.
Cause `ranger` is not really installed into the brewed python, but
running as a python file by it. To make sure pkg `ranger` being found
in `sys.path`. We must define a helper under "#{libexec}", which is
the working directory and pkg source `ranger/` is stored there.