cantata 2.4.2 (new formula)
As discussed in Homebrew/discussions#1370 and CDrummond/cantata#1671,
the Cantata maintainer is no longer providing precompiled binaries for
Cantata releases. Therefore, the `cantata`
cask (https://formulae.brew.sh/cask/cantata) cannot be updated beyond
`v2.3.2`, which was released in August 2018.

This adds a core formula to supercede the cask distribution. The
formula builds a `.app` bundle, and the user is informed via caveats
that they may wish to link it to their system directory. A symlink to
the `cantata` binary is also exposed directly in the formula's `bin/`
directory.

See #3331 for precedent for this sort of migration.

Co-authored-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>