libimobiledevice-glue: Added new package
This commit adds the (currently HEAD-only) package required to build
future versions of multiple libimobiledevice packages.
As the release cycle of libimobiledevice packages is slow, this package
is currently only required for HEAD-builds, but when the devs release a
versioned release of this package and update one of the dependencies,
this package is needed by at least 3 core packages at the time of
writing this commit (libimobiledevice, libirecovery, libusbmuxd).