libimobiledevice: Added dependency and fixed license
This commit adds the dependency "libimobiledevice-glue" required in
future releases of this package. As the current version does not use
this dependency yet, it has been added to the head section with a
reminder comment to move it to be a main dependency in the future.
It also adds the missing "-or-later" suffix to the license string.
The missing 'branch: "master"' attribute has been added to the
head URL, too.