envoy: remove linux build
I don't expect this change to be universally popular. However, I think
it is the most pragmatic way to isolate the impact of envoy on
maintainers and end users. Specifically, this removes the linux build
and in doing so removes fragile and complex configuration. It also will
reduce the time to land new versions on OS/x.

I believe this is justified beyond the technical dept recovered. Envoy
is not often installed on linux with homebrew. To the degree it can be
used, it can't be used on arm64 which is a supported Envoy platform on
Linux. The few using linuxbrew have other options including the official
docker images, func-e, and in the future, formal system packages and
tarballs.

Once we remove this, maintaining homebrew envoy's primary audience, OS/x
is easier and faster. We can focus energy on what's not available
elsewhere: a macOS build of Envoy. We can focus on the larger audience
which extends larger still when Envoy completes arm64+osx (likely in
version 1.21). Maintainers do not have to run very long linux builds
anymore, and CI resources can be conserved for projects more often used.

More details on the discussion are here, but I'm raising this PR as a
call to action. If the homebrew leads decide it is better to keep trying
to build on Linux, the main impact is continued longer and more complex
work to release new versions to end users. That's not the worst outcome,
but I believe they deserve better, hence this issue.

See https://github.com/Homebrew/discussions/discussions/2271

Signed-off-by: Adrian Cole <adrian@tetrate.io>