ldc: add comment explaining `test` block workaround
Setting `fails_with :gcc` will make `brew` set `CC=llvm_clang` during
the test block, but `llvm_clang` is a shim that won't be in `PATH` and
shouldn't be used for the test.

The workaround can look a little cryptic, so let's add a comment for a
little bit of context to explain what's going on here.