flat_namespace_allowlist: add `mhash`
The Catalina bottles have also been built with the `-flat_namespace`
flag, so there is no bug in detection macOS 11+ in their `configure`
script.

That said, they seem to be using a very old version of the snippet that
generates macOS linker flags, so we'll probably want to report this one
upstream too.