flat_namespace_allowlist: add `omniorb`
Use of the `-flat_namespace` flag does not come from the Libtool bug,
and omniORB's build system is a bit too complex for me to be comfortable
trying a patch.

I've reported this issue via email to omniORB-dev and omniORB-list. The
omniORB-dev archives don't seem to be updating, but my email should
hopefully appear at the omniORB-list archives [1] soon.

[1] https://www.omniorb-support.com/pipermail/omniorb-list/