git-ftp: de-vendor `curl`
The vendored `curl` is woefully out of date, and should use the formula.

This was added in 431fec3f9bfde190289b61c6f26db07c80d4c187 because
`git-ftp` needed a version of `curl` compiled with `libssh2`, and it was
unclear whether the `--with-libssh2` option would survive the removal of
options.

Conveniently, our `curl` is compiled with `libssh2`, so this formula can
just depend on our `curl` without any further changes.

This enables bottling on Monterey, because the vendored `curl` has a
`configure` script bitten by the `-flat_namespace` bug.