neovim: vendor `libvterm`
Neovim is broken with the Libvterm bump to 0.2 (#88610). A fix is in
progress upstream (neovim/neovim#16219), but it looks like sorting out
the regressions will take some time.

In the meantime, let's vendor and statically link the correct version of
Libvterm to fix the breakage that users face right now.

Closes #88665.