aom: align macOS and Linux builds
`jpeg-xl` is now bottled on Linux. I've omitted the `revision` bump
since it's not clear Linux users are looking for these new features, so
it doesn't seem necessary to force everyone to reinstall this. We can
always bump the revision later if we want to make sure this change
reaches more users.