ncurses: fix build failure + improvements
`configure` gets confused by a non-existent `lib/"pkgconfig"`, so we
create it before invoking `configure`.

While we're here:
- also build the C++ shared library
- reduce duplication in `make_libncurses_symlinks`
- also install an `ncurses` symlink in `include`

Incidentally, the Ncurses maintainer explicitly discourages symlinking
the w-suffixed libraries to the unsuffixed libraries because these two
are not ABI-compatible. It seems to work, so perhaps formulae that use
Ncurses might be mistakenly linking with `-lncurses` when they actually
want `-lncursesw`.