tests: set timeout via flags only on self-hosted runners
We're not particularly constrained (yet?) in our use of GitHub runners,
so we probably don't need to cut jobs short when using them. This would
be useful for PRs such as #89617 where the macOS runners finish in under
an hour, but the Linux runner requires more than an hour to complete.