pango: fix Meson flags
`use_fontconfig` is a non-existent option, so attempting to build this
with the latest version of Meson fails. This has been replaced with the
`feature`-type option `fontconfig`, which can be passed one of `auto`,
`enabled`, or `disabled`.

While we're here, let's make the dependency on `freetype` explicit (this
is a dependency of `fontconfig`, so this adds no new dependencies to
current installations), and make sure that the build doesn't silently
ignore failures in finding dependencies.