bug.yml: fix missing `label` property
Submitting bug reports hasn't worked for a while because of the missing
`label`.

My editor complained about `labels` needing to be an array, so I've made
that one too. It was also complaining about the missing label property
for the checkboxes, so I guess it knows what it's talking about here.

See https://github.com/Homebrew/discussions/discussions/2526#discussioncomment-1731168