libvirt: Tweak URLs
The canonical URL to the libvirt homepage doesn't contain the
www part, and in fact if you try using that version you will
get immediately redirected.

The primary git hosting for libvirt is GitLab, with GitHub
only being used as a mirror, so it makes sense for head to
point to the former.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>