Update Dungeon formula to work on Apple Silicon
I just got an M1 MacBook so I finally managed to troubleshoot why this formula wasn't building on Apple Silicon. The `opt_pkgshare` path on Arm was so long it overran column 72 and made gfortran flame out. Installing the files to `opt_share` instead seems to fix the problem on my machine.