yarn: make `node` a test dependency
Yarn can be used with multiple versions of Node, but depending on `node`
specifically breaks that.

This change allows `yarn` to be used as a dependency for
`opensearch-dashboards` (cf. #90403). See also yarnpkg/website#913.

While we're here, let's adjust the `NPM_CONFIG_PYTHON` value, since the
`node` formula doesn't even depend on a Python3 formula at runtime
anymore (and it's not that picky about the Python3 version anyway).