neovim: don't clobber default `LUA_PATH` and `LUA_CPATH`
The way we set `LUA_PATH` and `LUA_CPATH` makes it so that the default
search path is completely ignored, which can lead to build failures
(e.g. neovim/neovim#16688). We fix this by appending `;` to each
variable to make sure that the default paths are searched after the
directories we set.

While we're here, let's make sure we use the right path separator (`;`,
not `:`).

Fixes neovim/neovim#16688. (Actually, that has already been fixed by
neovim/neovim#16714, but the upstream fix has the undesirable consequence
of disabling pre-compilation of LuaJIT bytecode. This change re-enables
pre-compilation.)