dosbox-staging: downgrade to 0.78.0
`dosbox-staging` was updated to 0.78.1 because upstream had published
a `v0.78.1` tag (since removed). This used the stable tag format but
there was no corresponding release on GitHub, no announcement on the
homepage, etc.

Upstream confirmed this wasn't a release version and `v0.78.1-rc`
would have been a more appropriate tag. It's unclear when they will
release a new version, so this commit downgrades the formula to the
latest stable version, `0.78.0`.

This also adds a `livecheck` block using the `GithubLatest` strategy,
to align with the upstream release process (i.e., a version tag isn't
a sufficient indicator of a new release and we need to check GitHub
releases instead).