julia: switch to LLVM 13
Julia now builds against LLVM 13 by default on `master`, so it's
probably (relatively) safe to depend on LLVM 13 instead of LLVM 12 now.

This requires applying patches that have already been merged upstream. A
couple of them need to be backported, but thankfully Arch have already
done that, so we can just borrow the patches that they're using.

While we're here, let's make sure that the curl patch is fetched from a
stable URL to avoid checksum invalidation.