sc-im: default to native clipboard on macOS
The Makefile has the option to use `pbcopy` and `pbpaste` as the default
clipboard commands, but leaves them commented out with the suggestion to
uncomment these lines if you want to use them. Let's do that on macOS
since the alternative clipboard commands (`tmux` and `xclip`) require
additional dependencies.

While we're here, let's also adjust the maximum number of allowed rows
to the largest possible value. The default is set so that this can run
on devices with limited memory, but that's not really an issue for us.

We also define `-DGNUPLOT` to enable use of `gnuplot` for plotting if
available on the user's system.

Other package managers (e.g. Termux, MacPorts) make the same changes to
their build.