aerc: Set SHAREDIR to symlink rather than {prefix}
This commit changes the way aerc is built from source to enable
seamless upgrades on the same machine.

When aerc is run for the first time, a preferences file is written to
~/Library/Preferences/aerc.conf that includes references to the Cellar
directory that Homebrew maintains. This breaks after an update because
the directory is versioned and still references the old version.

This commit instead makes aerc use the symlink at /usr/local/share
which is also maintained by Homebrew and always points at the current
version installed by the user.

Fixes: #93407