rarian: regenerate `configure`
This is affected by the `-flat_namespace` bug from Libtool, but none of
our patches apply. Let's regenerate the `configure` script instead.

This is needed for bottling on Monterey.

While we're here, let's add a `test` block and specify the `license`.
The license data was taken from Debian's copyright file [0].

[0] https://metadata.ftp-master.debian.org/changelogs//main/r/rarian/rarian_0.8.1-6_copyright