rpm: fix build on Apple Silicon
Add two upstream patches to the RPM formula. The first addresses the
usage of *64 stat symbols, which have been removed on the Apple Silicon
toolchain. The second patch removes a broken Lua regular expression
extension. This extension has been broken for fifteen years, and
references a symbol that has been removed in recent versions of Lua.

This PR also removes the generic libtools configure script patch
for fixing the flat namespace bug. This is not necessary, as we regenerate
the entire configure script due to the Lua patch touching a `Makefile.am`
file.

This PR also changes the `inreplace` in `post_install` with specififying
`__GPG` as an argument to `configure`.