workflows/tests: always upload logs
I think the logs of successful runs are often still interesting (e.g.
for comparison of a successful run on one runner vs a failed run on
another). Moreover, given that we no longer fail CI when a formula
doesn't have a bottle, not all successful CI runs are successful builds,
so the logs may still be of interest even when CI is successful.