versioned_dependencies_conflicts_allowlist: add `freedink`
This is needed for #94505. The conflict comes from `freedink` depending
on `python@3.10` and `python@3.9` through `cxxtest` and `glib`
respectively.

`cxxtest` is installed into a virtualenv in `libexec`, so it's unlikely
that `freedink` actually makes use of anything there since it ships no
Python code itself.

An alternative to this is to migrate `glib` to `python@3.10`, which will
take quite a while (see #87277), or migrate `cxxtest` back to
`python@3.9`, which is unnecessary.

I've tested `freedink` locally and it seems to work fine despite the
mixed recursive dependency.

See also #65831.

Closes #94505.

Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>