workflows/tests: error when `long build` is set without `CI-long-timeout`
We now use a `long build` label to identify PRs that need the
`CI-long-timeout` label but don't have it because other PRs are
currently using it.

Let's save some CI time by failing CI whenever a PR has the `long build`
label but not the `CI-long-timeout` label.

I've also updated `triage.yml` to add the `long build` label where
appropriate. The criterion I used here was whether a CI run that skips
dependent testing will still exceed the short timeout of 60 minutes.

This leaves out formulae that regularly take a long time only due to
having a large number of dependents (e.g. `python@3.*`, `go`). I think
this is the correct thing to do to avoid having to manually remove the
`long build` label every time the `triage` workflow is run in cases
where the CI run will not actually need `CI-long-timeout` (e.g. when we
use `CI-skip-dependents`).

This is based on the suggestion from #94792.