workflows/tests: always test dependents unless requested otherwise
Skipping testing dependents makes it that much harder to identify
formulae that need revision bumps.

See, for example, #91224, where the need to rebuild `gnuradio` could've
been identified days, if not weeks, earlier if we did not skip dependent
tests when one of the tested formulae fails.

Note that this will still skip the testing of dependents of failed
formulae, because these are passed in the `--skipped-or-failed-formulae`
flag.

Related discussion: #82220