cpu_features: build shared library
Linux `ld` doesn't like linking with the static library because it
wasn't built with `-fPIC`. Let's fix that by linking with the shared
library instead.

We continue installing the static library to avoid breaking users who
expect it to be there (because that used to be what was installed).

Unblocks #95398.