mold: fix `inreplace`, improve test
Using `*` is greedy, so `(?:ld64\.)?` was not matching anything. We can
fix this by using the lazy `.+?` instead.

While we're here, let's run some of these tests that we're installing in
the `test` block. They don't take long, and are actually the kinds of
tests that we like (e.g. try to link some things using `mold` and
checking what's produced).

I've skipped doing the same on Linux since most of the tests require a
`clang` installation, and I'd rather not carry around an extra test
dependency here.