nnn: use ncurses from Homebrew
The ncurses that ships with macOS is quite old and does not contain a
terminfo entry for tmux-256color. Even if you manually compile the
tmux-256color terminfo entry from Homebrew, the native ncurses still
won't work because of a known bug that misinterprets the 'pairs' entry
[1]. Instead, use ncurses from Homebrew which ensures nnn works properly
in tmux out of the box.

[1]: https://github.com/tmux/tmux/issues/2262#issuecomment-719985206