Use AppleSpell by default on macs.
Enchant always reads its system-wide enchant.ordering file, even if the user file (~/.config/enchant/enchant.ordering) exists.  A user file with just "*:AppleSpell" in it overrides the "*" entry in the system-wide file but not the entries for enAU, enGB, etc., which are there by default.  As a result, the automatic language detection of AppleSpell does not work if the system language is one of those (enAU, etc.).  This change makes the system-wide file contain only "*:AppleSpell", so that spell checking with enchant works the same as in other mac apps.