Remove autobump formulae with Repology issues (#94346)
The newest version reported by Repology for these formulae is
inappropriate with respect to the formula. Since the `autobump`
workflow is now using `brew bump` to open version bump PRs and `bump`
uses the newest version from livecheck and Repology, we have to remove
these formulae from this list until we can tell `bump` to only open a
PR for a new version from livecheck.