spdlog: fix macro redefinition warning
When configuring spdlog with -DSPDLOG_FMT_EXTERNAL=ON, the generated
spdlog.pc file contains -DSPDLOG_FMT_EXTERNAL in its CFlags. This
means that when tweakme.h also contains #define SPDLOG_FMT_EXTERNAL,
it triggers a macro redefinition warning. This commit changes
tweakme.h to only define the macro if it's not already defined.