monkeysphere: Replace dead URIs
The `homepage` currently referenced in the formula points to a shady-looking
domain parking website (which, at the moment, also happens to have an expired
HTTPS certificate). Notably, the current `homepage` appears to have nothing
to do with monkeysphere. The closest thing I can find to a new homepage is the
debian tracker entry for the monkeysphere package.

Also, the repository referenced by `head` in the formula no longer exists, so
I have deleted that.