imagemagick: Use LibRaw instead of ufraw
ufraw is essentially abandonware at this point, and is missing support for many
newer cameras. This results in incorrect color interpretation in the best of cases,
and sometimes garbage output in the worst scenarios. LibRaw is more modern, and is
actively maintained. It has support for newer cameras, interpets colors more
consistently, and is also quite a bit faster.

Closes #98195.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>