R: Remember path to gfortran
R constructs a "Makeconf" file at R's build time that remembers paths to and flags for compilers and other important utilities.
The same compilers and settings are used to build CRAN packages when end users call install.package().

If R finds gfortran in PATH at build time (and, as Homebrew builds it, it does), it will remember the value "gfortran".
CRAN package builds will also try to invoke "gfortran", which will end up searching PATH.
If Homebrew's prefix/bin is not in PATH, these builds will fail.

Homebrew can be missing from PATH if R is running under an IDE like Rstudio, and does not inherit environment variables set in shell initialization scripts.

This change encourages R to remember the absolute path to the Fortran compiler in Makeconf so that Homebrew does not need to be in PATH.