zstd: use shared linkage, remove `zlib` dependency
This formula has no linkage with `zlib` on macOS and Linux, so it seems
likely that `zlib` is not actually needed.

Additionally, set `ZSTD_PROGRAMS_LINK_SHARED` so that the `zstd` binary
links dynamically with `libzstd` instead of linking statically.

Shared linking decreases the size of the `zstd` binary and creates
opportunities for more efficient use of system resources (at the minor
cost of overhead care of dynamic linking).

Neither of these changes are probably enough to warrant a revision bump,
but building new bottles won't hurt.