synergy-core 1.14.4.37 (new formula)
Synergy is a free and open source program for sharing a mouse and keyboard
among multiple computers. It supports 64-bit Intel Mac, 64-bit ARM Mac,
64-bit GNU/Linux, and other platforms.

Synergy was previously available in brew as a cask named `synergy`. That
cask installed binaries published by Symless, the corporate sponsor of
Synergy. However, at some point in the last couple months, Symless decided
to start charging money for access to the binaries on its website. As part
of this change, the Symless binaries were placed behind an authentication
webpage. As a result, the `synergy` cask was removed from brew:
  https://github.com/Homebrew/homebrew-cask/commit/0037f2409

However, even though Symless no longer publishes binaries, Synergy remains
fully free and open source software, licensed under GPL version 2 (with an
exception allowing "compiling, linking, and/or using OpenSSL"). As a result,
nothing stops us from building our own binaries from the source code. This
new `synergy-core` formula does exactly that: it builds Synergy from source
and installs it.

This formula supports 64-bit Intel Mac, 64-bit ARM Mac, and 64-bit GNU/Linux.
I have tested the formula on all three platforms. (For GNU/Linux testing, I
used Fedora 35, although that probably doesn't matter.)

This formula sets up a `synergy-core` daemon that can be run and managed
using `brew services`. This daemon is currently how I recommend running
the binaries built by this formula.

***

The `synergy-core` project is distributed under the GPL-2.0 license with an
exception that grants additional rights to the user. The project's LICENSE
file reads as follows:
  This program is released under the GPL with the additional exemption
  that compiling, linking, and/or using OpenSSL is allowed.
This preamble is followed by the text of the GPL-2.0.

This is a free software license but it cannot be represented with the `brew`
license statement, so the formula uses `license :cannot_represent`.
Unfortunately, the GitHub Licenses API incorrectly states that `synergy-core`
is licensed strictly under the GPL-2.0. So we need to add `synergy-core` to
  audit_exceptions/permitted_formula_license_mismatches.json
to avoid `brew audit` objecting that the license specified in the
`license` statement is different from the license returned by the
GitHub Licenses API.

Co-authored-by: Sean Molenaar <SMillerDev@users.noreply.github.com>
Co-authored-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>