minimal-racket: run raco setup post install
Racket relies on the mtimes of compiled files (.zo) to be newer than
source files (.rkt), otherwise it falls back to interpreting. For core
libraries, this is the difference between a 200ms and 15s startup time.
Multiple users have reported these extremely slow startup times after
installing from Homebrew, which is fixed by running `raco setup`. This
change therefore runs that in the post_install step. When nothing needs
to be recompiled, `raco setup` completes in a couple seconds.

See this discussion for more details:
https://racket.discourse.group/t/minimal-racket-is-very-slow-before-installing-compiler-lib/967