Add new formlua: wasi-sdk (15.0).
WASI-SDK is a distribution of a custom libc (wasi-libc) with clang/LLVM
to target the WASI (WebAssembly System Interface) platform. WASI is a
definition of a POSIX-like set of syscalls for WebAssembly modules to
use. A program compiled for the `wasm32-wasi` target can run in various
WebAssembly-outside-the-browser environments (server-side, cloud/edge
computing, command-line runtimes like Wasmtime, etc) and within browsers
with the appropriate shims.

This formula is a fairly thin wrapper around the custom Makefile in the
`wasi-sdk` repository [1], which itself is a fairly thin wrapper around
an LLVM/Clang build with specific options and defaults together with a
build of the custom sysroot.

[1] https://github.com/WebAssembly/wasi-sdk/