libmarpa: update HEAD build procedure
Note: this passes the  `brew audit` _except_ for this message, which looks like a bug in brew(?):
```
libmarpa:
  * HEAD: Use `branch: "master"` to specify the default branch
Error: 1 problem in 1 formula detected
```

The branch is being changed because the repository's default branch (master) is intended to be in a broken state, in general.  The maintainer established the "tested" branch for states that work.