justprep 1.2.3 (new formula)
The first commit from MadBomber for the justprep
brew formula which does not contain bottles.

This formula builds the Crystal implementation.

There are two implementations: Ruby and Crystal.  The Ruby version is released through the rubygems.org.  The Crystal version is released through homebrew.

justprep allows a "justfile" to be auto-generated from a seperate source file that contains inclusionary keywords such as include, import, require and/or with. These keywords are followed by a path to a file. If the file exists, the contents of the file are inserted into the justfile at the position of the keyword command. After all keywords have been process a new "justfile" is created which can then be used by the 'just' command-line tool.