terminalimageviewer: remove `assert_equal`
The output seems to depend on the version of `imagemagick`, which makes
this an unreliable test. Even if it didn't, the output is gibberish for
anyone who isn't a terminal, so I don't think there's much value in
checking for it.

Let's just make sure we can run `tiv` without any errors instead.

Needed for #105299.