openconnect: improve `vpnc-script` handling
1. The default location of this script is `/etc/vpnc`, so let's use the
   analogous location in `HOMEBREW_PREFIX`.
2. Add caveats that flag the install location of the vpnc-script and
   warn the user when a newer version exists.

The change in location has the added benefit of switching users to a
newer version of the vpnc-script automatically, since there have been
recent updates to this script.

Also, let's make the use of `pipe_output` instead of `shell_output` in
the `test` block clearer by beign explicit about passing an empty string
and adding a comment.