workflows/tests: show job summaries of each runner
This will display some results of the CI run in the workflow run summary
page, so you won't need to click on individual jobs to see what happened
there.

See https://github.blog/2022-05-09-supercharging-github-actions-with-job-summaries/

There's probably room for improvement in what exactly we put in
`$GITHUB_STEP_SUMMARY`, but that's something we can try later on.