libphonenumber: use Homebrew `abseil`
This downloads its own `abseil` during the build, and then gets confused
about where to find `abseil` headers when you try to link with it.

Let's fix that by relying on our version of `abseil` instead.

This should help unblock #105712 and previous attempts to bump this
formula's version (e.g. #105188, #101703, #98994, #96834).