zstd: enable zlib/lzma/lz4 support
Many major distros (e.g. Debian, Fedora, Arch, MacPorts) enable this in
their zstd package, and it would be nice to get the same `zstd` whether
you installed it with `apt` or with `brew`.

Also, it's a huge QOL improvement to just be able to do

    zstd -d foo.{gz,lz4,lzma,xz,zst}

instead of having to work out which decompression utility to use
depending on the file format.

Let's also update the test to make sure that `zstd` can compress to and
decompress from the various applicable formats.