elixir-ls: Use compiled builds
Right now `elixir-ls` build itself from source with `elixir` dependency installed from brew.

The problem is that it causes to use latest version of Erlang, build artifacts of which are not backward compatible (e.g. OTP-25 build won't run on anything but OTP-25).

Because of that, [it messes with `elixir-ls`' support of last 3 version of Elixir and Erlang](https://github.com/elixir-lsp/elixir-ls#supported-versions=) if you try to use it in projects that use different Elixir/Erlang version, e.g. via asdf (which `elixir-ls` explicitly supports).

This PR uses precompiled releases of `elixir-ls` built **with minimum
versions of Erlang and Elixir supported** by the project, which makes it
forward compatible with their newer releases.