pdfpc: fix `HEAD` build
The latest version of `pdfpc` (v4.5.0) does not build on macOS because
it requires `webkitgtk` [1, 2], which does not build on macOS [3, 4].

Conveniently, `webkitgtk` is now optional on `master` [5, 6]. However,
building `HEAD` also requires a few extra dependencies. Let's try to fix
that so that users who want the latest version can try to build from
`HEAD` (at least until the next time it breaks).

Inconveniently, this introduces a mixed version dependency, since
`pdfpc` requires `libsoup@2` and `gst-plugins-good`, which depends on
`libsoup`. But, for now, this just on `HEAD` so it's not something we
have to worry about. We can disable either the `MOVIES` feature
(requires `gst-plugins-good`) or the `REST` feature (requires
`libsoup@2`) in the event that we need to resolve the dependency
conflict.

Closes #106759.

[1] https://github.com/pdfpc/pdfpc/blob/0805bd3b2eb7a2f0d14ed045c81fbc298c4202d1/src/CMakeLists.txt#L7
[2] https://github.com/Homebrew/homebrew-core/pull/67319
[3] https://github.com/Homebrew/homebrew-core/pull/11243
[4] https://github.com/Homebrew/homebrew-core/pull/104779
[5] https://github.com/pdfpc/pdfpc/blob/3310efbf87b5457cbff49076447fcf5f822c2269/src/CMakeLists.txt#L32-L36
[6] https://github.com/pdfpc/pdfpc/issues/590