lasso: depend on `libxml2`
This already links with our `libxml2` indirectly through `libxmlsec1`.
This means, in particular, that we should not be mixing the system
`libxml2` headers with the library that we ship.

I've also removed the call to `ENV.prepend_create_path` as this just
creates an empty directory in `libexec`.

Let's also build this using macOS `python3` where we can.