graphicsmagick: disable jpeg-xl support, which seems to require unreleased version of libjxl
Building graphicsmagick 1.38 locally fails in `coders/jxl.c`.

The failure in the build log starts with the following error:

```
coders/jxl.c:811:3: error: use of undeclared identifier 'JxlEncoderFrameSettings'
  JxlEncoderFrameSettings
  ^
```

It seems version 0.6.1 of jpeg-xl doesn't really support this
JxlEncoderFrameSettings symbol, it was introduced in order to deprecate
JxlEncoderOptions, but that will only be available in version 0.7 of jpeg-xl,
which is not yet released. In particular, the release notes for the release
candidate for 0.7 mentions this specific change, among others:

https://github.com/libjxl/libjxl/releases/tag/v0.7rc

So it turns out the code in graphicsmagick 1.38 depends on jpeg-xl 0.7, which
is not available yet.

Furthermore, the `--with-jxl` flag was turned on by default just recently, as
version 1.38 of graphicsmagick was released.

https://sourceforge.net/p/graphicsmagick/code/ci/b2a4fb7af59b74ddb3bc745b0d1db3022d7437c8/

Therefore, disabling it explicitly here doesn't look like a bad idea.

When jpeg-xl 0.7 is released and the formula for it is updated, we can go back
to re-enabling jpeg-xl support here again.

Tested: `brew reinstall -s graphicsmagick`