podman: Build a relocatable binary, using upstream feature to locate helpers relative to the main binary
PR containers/common #1127 (https://github.com/containers/common/pull/1127)
introduced a new feature by which podman is now capable of locating its helpers
(such as `gvproxy`) relative to the main binary location, and by default
enables that feature for MacOS builds.

Apply the patch on top of the current Homebrew tree and enable the feature.

Move `gvproxy` from a bare `libexec` to a `libexec/podman`, since that's the
default path to the binary used in podman.

For Linux builds, use the environment variable consumed by the Makefile to
inject the same binary-relative path to the default helper search list.

Since this change makes this keg relocatable again, drop the
`pour_bottle? only_if:` gate that was causing it to build from source
whenever using a non-standard prefix. Bump up revision number. Installing from
bottles should now work fine, even on a non-standard prefix.

Tested: Rebuilt from source with `brew reinstall -s podman` and confirmed that
`podman machine start` works (that's the operation that depends on the location
of the `gvproxy` binary.) Tested a different directory location to test (on a
Mac) that the Linux change should work as exepcted.