pandoc-crossref: revision for pandoc 2.19.2 update
`pandoc-crossref` warns when it was compiled against an older version of `pandoc` than its current invoker is. This avoids that warning and any weirdness resulting from the version mismatch. E.g.

```
WARNING: pandoc-crossref was compiled with pandoc 2.18 but is being run through 2.19.2. This is not supported. Strange things may (and likely will) happen silently.
```