readline: remove unnecessary `inreplace`.
Let's also simplify the `install` method a bit.

The `inreplace` was needed so that `readline.pc` did not declare

    Requires.private: termcap

However, the correct way to avoid that is to pass `--with-curses` so
that it instead does

    Requires.private: ncurses

which allows `readline.pc` to be usable without modification.

Passing `--with-curses` does not change the configuration on macOS. The
existing bottles already have linkage with the system `ncurses` library,
because linking with `ncurses` is the default on macOS:

    http://git.savannah.gnu.org/cgit/readline.git/tree/support/shobj-conf?id=5263c0d88064fda96292335d12eec1733f91cdc9#n174

We can use this fact to avoid having to condition passing `SHLIB_LIBS`
on Linux, because this flag has no effect on macOS. Arguably, we should
not be setting `SHLIB_LIBS` at all, the documentation and source both
suggest that readline shouldn't be linking with `ncurses` on Linux.
However, removing this flag is a more substantial change that I'd like
to leave for another patch.