speedtest-cli: fix test
Test fails with: ERROR: Unable to connect to servers to test latency.
I can reproduce this locally too, maybe upstreams servers are down / something has changed.
Note that I am on tethered connection on a phone right now, so this might explain the issue on my side.

Testing the version and help commands is probably enough here as we just make sure the python code
can be executed (there is only a single python file in this project), and avoids any fragile
connection to speedtest's servers