igraph: suite-sparse is not a dependency any more
igraph 0.9.11 and earlier versions required suite-sparse as
a dependency. From igraph 0.10, suite-sparse is not a dependency any
more, but the Homebrew formula still declares it as a dependency
incorrectly. This PR fixes the issue by removing suite-sparse from the
list of dependencies.

igraph 0.10 includes a vendored copy of CXSparse, a sparse matrix
library that is part of suite-sparse, but it is patched to ensure that
the size of integers in CXSparse matches the size of integers in igraph,
even if igraph is forcibly compiled with 32-bit integers on 64-bit
platforms. Therefore, a stock suite-sparse is not suitable for linking
to igraph.