sox: fix device name length
This commit adds a patch created in 2016 which fixes a very annoying bug in macOS sox.  The bug causes the display of device names to be pinned to the length of the last device. Please see 'https://public-inbox.org/sox-devel/1465929413.3110479.637612809.34A344C0@webmail.messagingengine.com/T/#u' for the sox-devel mailing list discussion.