Add the `licensure 0.2.0` formula
`licensure` is a program that can generate and validate license and
copyright comment headers across one's project sources.

The latest tagged version is 0.2.0, but as of now, this tag fails
to build. This issue was fixed by the immediate-next commit, though
the author (@chasinglogic) did not release another tag with this fix.
Thus, in this formula, we apply the same fix as an embedded patch.