asdf: Remove placeholders
We're past August 31st, 2022!

AFAICT, these placeholders never worked quite right, and have caused an
*insane* amount of confusion over on the asdf issue tracker. See
https://github.com/asdf-vm/asdf/issues/1103#issuecomment-1093094933 for
a writeup explaining how using these placeholders results in shims that
break the next time you do an asdf upgrade.

You may note that the TODO made it look like we should remove the
`bin.write_exec_script libexec/"bin/asdf"` as well. I left that in
because without it, we do not end up with an asdf binary in our path at
all. I don't know enough about Homebrew to know if this is actually the
right way to get a binary on the PATH or not.

Before
======

    jeremy@ homebrew-core % ls "$(brew --prefix asdf)/asdf.sh"
    /opt/homebrew/opt/asdf/asdf.sh
    jeremy@ homebrew-core % ls "$(brew --prefix asdf)/libexec/asdf.sh"
    /opt/homebrew/opt/asdf/libexec/asdf.sh

After
=====

    jeremy@ homebrew-core % ls "$(brew --prefix asdf)/asdf.sh"
    ls: /opt/homebrew/opt/asdf/asdf.sh: No such file or directory
    jeremy@ homebrew-core % ls "$(brew --prefix asdf)/libexec/asdf.sh"
    /opt/homebrew/opt/asdf/libexec/asdf.sh

- [x] Have you followed the [guidelines for contributing](https://github.com/Homebrew/homebrew-core/blob/HEAD/CONTRIBUTING.md)?
- [x] Have you ensured that your commits follow the [commit style guide](https://docs.brew.sh/Formula-Cookbook#commit)?
- [x] Have you checked that there aren't other open [pull requests](https://github.com/Homebrew/homebrew-core/pulls) for the same formula update/change?
- [x] Have you built your formula locally with `brew install --build-from-source <formula>`, where `<formula>` is the name of the formula you're submitting?
- [x] Is your test running fine `brew test <formula>`, where `<formula>` is the name of the formula you're submitting?
- [x] Does your build pass `brew audit --strict <formula>` (after doing `brew install --build-from-source <formula>`)? If this is a new formula, does it pass `brew audit --new <formula>`?