dynare: remove Linux bottle
The Linux bottle is likely broken, because the test fails reliably (i.e.
100% of the time) on Linux. The constant test failures create lots of
noise in CI, and force us to merge PRs even when CI is failing.

To try to minimise the number of PRs merged with failing CI, let's just
remove the bottle.

On a related note: while the Dynare test has had a history of flakiness,
I don't think I've seen a random failure for it on macOS recently.

Closes #110308.