pythran: update test
The second test case is broken on Linux until we rebuild `pythran` using
the newer GCC. This should happen when `python@3.10` is version 3.10.7
or newer.

Let's stop skipping the test on Linux when `python@3.10` has been
rebuilt, but throw an error as a reminder to remove the hacks in the
test when the time comes.