rgbds 0.6.0
Update RGBDS to v0.6.0, and rgbobj to v0.2.1.
This also fixes the clap dependency issue, therefore the patch is no longer necessary.

The build system was switched to CMake for better cross-system compatibility.
The Makefile currently passes `-flto=auto`, which is not supported by Clang before 13,
and thus fails to build on macOS before Monterey.

Additionally, add a test to check that RGBGFX correctly runs.