Test formula, please let CI run but do not merge into Homebrew
This PR is submitted to test the behaviour of Python 3 during the build of a
formula. Submitting a test / pre-release PR was suggested by @carlocab
in https://github.com/orgs/Homebrew/discussions/3773

So, could you please let this PR run its CI builds so that the behaviour
of Python 3 during the build is reported in the log?

Of course, this formula should be dropped afterwards and not merged into
homebrew-core.

Thanks.