mold: make bottle relocatable again
With the changes in rui314/mold@0d7a5a7, libdir
is again embedded in the binary. We may work
around this by setting `MOLD_LIBDIR` to an empty
string. This works because the search for
mold-wrapper.so will fallback to
../lib/mold/mold-wrapper.so when a MOLD_LIBDIR
lookup fails. The mold-wrapper.so test guarantees
that.

See also:
- bd0a7c0
- https://github.com/rui314/mold/blob/v1.7.0/elf/subprocess.cc