cloudflare-wrangler2 2.6.2 (new formula)
This adds version 2 of Cloudflare's Wrangler utility as a separate formula.

It exists in a separate repository, https://github.com/cloudflare/wrangler2,
and is a rewrite of the utility that isn't entirely backwards compatible with
version 1.

It installs a `wrangler2` binary so it doesn't conflict with the existing
`cloudflare-wrangler` formula.