qt: fix build in Python 3.11 (invalid mode: 'rU')
In Python 3.11, 'U' ("universal newline") is no longer accepted in
the file mode, having been deprecated in Python 3.3. The "universal
newline" is used by default when a file is open in text mode.

This commit removes the 'U' from the two (non-third-party) places
it is used.

See https://bugreports.qt.io/browse/QTBUG-108436