re2: build and install static library
e98420fea827155ca9f9f3f20c1830723a0cad6c switched to using cmake, but
cmake only builds the shared (.so) or static library (.a) depending on
whether `BUILD_SHARED_LIBS` is enabled.

To fix this, we run `cmake` twice: once to build the static library,
and once to build the shared library.