passenger: use Homebrew-provided nginx
Build logs reveal that `passenger` is still downloading an Nginx tarball
because no arguments were passed to `./bin/passenger-config` (see [^1]).
Ensure that we're using the one provided by the `nginx` formula. We also
need to rebuild the tarball with Gzip, and make sure it has the
top-level directory that `./bin/passenger-config` can recognize (details
in [^1]).

Nginx has defaulted to `pcre2` since release 1.21.5 (see [^2]), so with
this change, we will also make `passenger` depend on `pcre2` to avoid
undeclared linkage. The `pcre` dependency needs to be kept because
`passenger` itself is still using it (see [^3]).

[^1]: https://github.com/phusion/passenger/blob/release-6.0.17/src/ruby_supportlib/phusion_passenger/config/compile_nginx_engine_command.rb
[^2]: https://github.com/nginx/nginx/commit/c6fec0b027569a4e0b1d8aaee7dea0f2e4d6052b
[^3]: https://github.com/phusion/passenger/blob/release-6.0.17/src/ruby_supportlib/phusion_passenger/platform_info/pcre.rb