gnupg@2.2: fix broken indirect linkage
This formula depends on `gnutls`, which currently depends on `guile`.
The `guile` dependency will be removed from `gnutls` in #122955, but
that will result in broken linkage for this formula.

We can fix that by declaring a dependency on the indirectly linked
formula.

The new dependency is already present in this formula's dependency tree,
so this change (by itself), should be a no-op.