pachi: removed formula
following homebrew's decision to deprecate caffe there is currently no way to
accomodate a modern (dcnn) pachi build in homebrew. the current formula is
severely outdated and leads to a stunted nodcnn build, which at best misrepresents
what Pachi can do. users who have an interest in Pachi are advised to compile it
from source or use the released binaries for a full-featured build.