fb303: add TODO for removal from `synced_versions_formulae`
I'm remove this from `synced_versions_formulae.json` in #124347 since
upstream seem to be missing a tag. Doing this allows us to update the
other formulae in the meantime.

However, let's add a note to add this back once upstream resolves the
issue we filed about the missing tag. See facebook/fb303#34.