pygit2 1.11.1 (new formula)
We currently bundle this as a resource with multiple formulae. However,
this causes difficulty when updating `libgit2`, because we often have to
rebuild all those formulae with some `libgit2` updates.

Let's simplify maintenance of those formulae by adding `pygit2` as a
formula instead. We can then have those formulae depend on `pygit2` to
minimise the required rebuilds from updates of `libgit2`.