benerator 3.1.0 (new formula)
Benerator is a powerful tool for generating, obfuscating, and migrating data for development, testing, and training purposes.
This formula makes it easy for users to install Benerator-CE   on their systems, without having to worry about manual setup or configuration.
The formula includes all necessary dependencies and configuration files, and can be installed with a single command. Users can then start using Benerator-CE right away, without having to spend time on tedious setup tasks.
This is the initial version of the formula, and we plan to update it regularly as new versions of Benerator-CE are released.