workflows: publish approved PRs upon CI completion
This is based on one of @MikeMcQuaid's suggestions on Slack.

We add an `automerge.yml` workflow that is triggered by the
`workflow_run` event. We need this in order to escalate access of
`GITHUB_TOKEN` from the `pull_request` and `pull_request_review` events,
which can only have read access to the GitHub API.

The `automerge` workflow is triggered by two events:
- completion of a `CI` workflow on a pull request
- approval of a pull request

Because both of these events are required in order to merge, we allow
the workflow to wait for a while to check if the other event occurs.

The basic idea of this workflow was borrowed from the documentation on
`workflow_run` events [1].

[1] https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#workflow_run