Add formula for Sqitch
Based on the sqitchers/sqitch tap, it supports PostgreSQL, MySQL, and
SQLite out of the box, and Snowflake, Exasol, and Vertica with ODBC via
the installation of additional client libraries. Snowflake support could
use the `snowflake-snowsql` formula, but it's Cask only. The caveat
output points to a URL that describes how to install these additional
dependencies, which will be updated appropriately by
sqitchers/sqitch.org#13.

The install method installs all Perl modules in the Cellar directory,
which is useful for keeping everything together and avoiding conflicts
with other Perl applications, but means that they must all be
re-installed when upgrading to a new version.

Not supported: Oracle and Firebase, since there are no formulas for
Instant Client or the Firebase server (or clients). The URL emitted by
the Caveat points users who need to manage Oracle or Firebase databases
to the Sqitch tap.