prqlc: renamed from prql-compiler
This renames `prql-compiler` to `prqlc`. This change was made many months ago, but I'm only getting around to updating the taps now. `prql-compiler` is the name of the library rust crate.

I followed the instructions at https://docs.brew.sh/Rename-A-Formula.

When this merges, we'll update our documentation to point here; it's been pointing at our own tap for the past year.