thrift@0.16.0 0.16.0 (new formula)
Thrift has bumped their minimum Java version to Java 11. However, there are many libraries that use Thrift that depend on Java 8. One of these libraries includes Parquet. As such, Parquet is unable to upgrade to the latest version of Thrift, and an ongoing conversation is occurring here: apache/thrift#2785.

Homebrew only has 2 versions of Thrift available, 0.9 and 0.18.+, the latter of which is on Java 11. I'd like to re-introduce Thrift 0.16.0 into Homebrew to give developers that are mandated to use Java 8 applications a working version of Thrift.