sox with alsa on linux (update formula)
Up to now, on linux, sox has been built with support for oss only,
which however is no longer the standard for modern linux
systems. Therefore invoking play (e.g.) fails with an error. This
patch adds build option '--with-alsa' to build sox with support for
alsa in addition to oss. As a result play actually produces sound.