feat: update ncspot for linux pulseaudio backend
This commit addresses a bug I encountered when installing ncspot on
Ubuntu 22.04. The current formula uses the `portaudio` backend, which
appears to be incompatible with Linux. The ncspot repository recommends
compiling with the default `pulseaudio` backedn on Linux, so I
implemented this for Linux (and retained `portaudio` for Mac). Also
updated the tests accordingly