lxi-tools 2.6 fix glib error
This formula revision fixes an error introduced in the original formula:
GLib-GIO-ERROR **: No GSettings schemas are installed on the system.
We mistakenly removed the schemas to workaround an error in the brew CI.
This revision regenerates the schema post install.
This is simply a formula revision, not a patch release.