libtomcrypt: fixed dependencies for asymmetric cryptography
In HomeBrew, libtomcrypt had all asymmetric cryptography disabled
because no bignum mathematical library was set. Using RSA, ECC and
the like was impossible with libtomcrypt in a HomeBrew environment.
On other systems, it works.

Required fixes:
- Add the required compilation options to enable LTM and GMP descriptors.
- Build the test program on a separate command because of additional
  mathematical library requirement.
- Add dependencies on libtommath and gmp, the two supported bignum
  mathematical libraries.

Note: On Ubuntu, for instance, libtomcrypt depends on libtommath
and gmp. With HomeBrew, the two dependencies were missing.

$ apt depends libtomcrypt1
libtomcrypt1
  Depends: libc6 (>= 2.17)
  Depends: libgmp10 (>= 2:6.2.1+dfsg1)
  Depends: libtommath1 (>= 1.1.0)