meson: pin down to python 3.11
Python 3.11 is needed in order to build glib. The distutils deprecation
prevents glib from building against 3.12.

Meson does not use environment variables and will use the python that it
was configured against when building. For this reason it is necessary to
pin meson down to python 3.11 in order to fix the glib build.

Signed-off-by: David Aguilar <davvid@gmail.com>