libpqxx: Improve test
It seems that this `libpqxx` static library is somehow broken. Some
basic programs, like the one currently provided in the formula, work
fine, but certain components of the library seem to fail.

This PR is extending the current test to hopefully surface the error I'm
seeing.

See https://github.com/jtv/libpqxx/issues/739
Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>