Enable LZVN and LZFSE compression support in `afsctool`
LZFSE is an advanced and performant compression algorithm by Apple.
Without this changeset transparent filesystem compression with
`afsctool` is limited to old Zlib compression algorithms.

Turns out, the LZFSE library code is attached to `afsctool` GitHub repo
as a Git submodule. GitHub tarball doesn't include submodule content,
hence the feature is unavailable. (As a matter of fact, `git archive`
and all the tools I'm aware of similarly don't handle submodules in a
bare repo.)

The proposed change is to add the missing submodule tree as a resource.
The approach mirrors how such things are done for `v8` formula.
Unfortunately, one has to update the revision hash of submodule every
time the main application is updated.

Although there exists a `lzfse` formula, depending on it wouldn't work
as `afsctool` doesn't rely on dynamic linking. Moreover, it doesn't
link statically either: it just boldly vendors in its dependency.
One also can't assume `afsctool` and `lzfse` are developed and updated
in sync.

See the following in `CMakeLists.txt`:
```
    if(EXISTS "${CMAKE_CURRENT_LIST_DIR}/src/private/lzfse/CMakeLists.txt")
        message(STATUS "Enabling LZVN and (possibly) LZFSE support")
        add_definitions(-DHAS_LZVN)
        add_definitions(-DHAS_LZFSE)
        add_subdirectory(src/private/lzfse)
        include_directories(src/private/lzfse)
        set(HAS_LZFSE 1)
    else()
        message(WARNING "Not enabling LZVN/LZFSE support - did you check out the lzfse submodule?!")
    endif()
```

The test is rewritten:
* made compression sample larger and compressible
* added a case with explicit LZFSE compression
* added decompression tests with verification
* removed the ZFS test as it's false-positive the way it was